Use a common credential for both ballot parts
As a result, the ballot's serial number and the part's tag must be
taken into account when generating long vote-codes and when using the
security code's randomness extractor. Otherwise, the same values would
be produced if the security code is not enabled. Also, from now on the
credential_length will be treated as the number of base32 characters,
not the number of bytes.