Update API's authentication and permission implementation
Define specialized permission classes to be used by REST framework's viewsets,
instead of relying on Django's model permissions. The reasoning behind this is
that each permission is tied to a single API user, and specifying a permission
class on each action makes the intention more explicit.

Furthermore, the APIAuthNonce model is moved to the base package.